feat: extract modal to accept components
